+++ src/com/nixinova/input/Keys.java
@@ -27,6 +27,7 @@ public class Keys {
 	public static final int NUM_7 = KeyEvent.VK_7;
 	public static final int NUM_8 = KeyEvent.VK_8;
 	public static final int NUM_9 = KeyEvent.VK_9;
+	public static final int NUM_0 = KeyEvent.VK_0;
 
 	// Mouse
 	public static final int LCLICK = MouseEvent.BUTTON1;